Remove string gens in buche mode javascript
